Add `--always-call-handle-client-data` flag
This optional flag will cause the `handle_client_data()` function of
`HttpProxyBasePlugin` plugins to be called even when `self.upstream` is
not `None`. The motivation behind this change is that I needed to be
able to read all messages sent over a WebSocket, and not just the
upstream responses which are caught by `handle_upstream_chunk()`.